+++ frontend/src/lib/components/sidebar/MenuButton.svelte
@@ -36,7 +36,11 @@
 		color?: string | null
 		trigger?: MenubarMenuElements['trigger'] | undefined
 		href?: string | undefined
+		/** Applied to the button AND to the label, so text sizing reaches the label text. */
 		class?: string | undefined
+		/** Applied to the button alone — for anything that would be wrong behind the label,
+		 *  a background above all. */
+		buttonClass?: string | undefined
 		// Show a trailing chevron to signal the button opens a dropdown.
 		showChevron?: boolean
 		// Render the label with stronger weight/size (e.g. the workspace name).
@@ -44,6 +48,9 @@
 		// Drop the native `title` attributes — for callers that wrap the button
 		// in their own hover tooltip.
 		disableTitle?: boolean
+		// Accessible name when the visible label is absent or only shown some of
+		// the time, so the button stays announceable in every state.
+		ariaLabel?: string | undefined
 	}
 
 	let {
@@ -64,9 +71,11 @@
 		trigger = undefined,
 		href = undefined,
 		class: classNames = undefined,
+		buttonClass = undefined,
 		showChevron = false,
 		emphasizeLabel = false,
-		disableTitle = false
+		disableTitle = false,
+		ariaLabel = undefined
 	}: Props = $props()
 
 	let buttonRef: HTMLButtonElement | HTMLAnchorElement | undefined = $state(undefined)
@@ -110,10 +119,11 @@
 				sidebarClasses.hoverBg,
 				'transition-all relative',
 				sublabel ? 'h-10' : 'h-8',
-				classNames
+				classNames,
+				buttonClass
 			)}
 			use:conditionalMelt={trigger}
-			aria-label={label}
+			aria-label={ariaLabel ?? label}
 			title={isCollapsed || disableTitle ? undefined : label}
 			{...$trigger}
 		>
