+++ frontend/src/lib/components/sidebar/OperatorMenu.svelte
@@ -9,26 +9,35 @@
 		Sun,
 		Code2,
 		LayoutDashboard,
-		Building,
 		Calendar,
 		ServerCog,
 		Table2,
 		GraduationCap
 	} from 'lucide-svelte'
+	import SwitchWorkspaceSubmenu from './SwitchWorkspaceSubmenu.svelte'
+	import {
+		OPERATOR_MAIN_LINKS,
+		OPERATOR_SECONDARY_LINKS,
+		OPERATOR_TRIGGER_LINKS,
+		type OperatorMenuLink,
+		type OperatorTriggerLink
+	} from './operatorRoutes'
 	import { base } from '$lib/base'
+	import { page } from '$app/state'
 	import { TOUR_PARAM, TOUR_PARAM_VALUE } from '$lib/components/tutorials/operatorTour'
 
 	import MultiplayerMenu from './MultiplayerMenu.svelte'
 	import { Plus } from 'lucide-svelte'
 	import {
-		clearWorkspaceFromStorage,
 		enterpriseLicense,
 		superadmin,
 		usedTriggerKinds,
 		userWorkspaces,
+		workspaceColor,
 		workspaceStore
 	} from '$lib/stores'
 	import { twMerge } from 'tailwind-merge'
+	import { getContrastTextColor } from '$lib/utils'
 	import { USER_SETTINGS_HASH } from './settings'
 	import { logout } from '$lib/logoutKit'
 	import DarkModeObserver from '../DarkModeObserver.svelte'
@@ -39,6 +48,10 @@
 	import type { FavoriteKind } from './FavoriteMenu.svelte'
 	let darkMode: boolean = $state(false)
 	let showExtraTriggers = $state(false)
+	// The menu opens on hover, so this doubles as "the pointer is on the button":
+	// it slides the workspace name out beside the icon, and keeps it out once a
+	// click has pinned the menu open and the pointer has moved away.
+	let menuOpen = $state(false)
 
 	interface Props {
 		isCollapsed?: boolean
@@ -51,104 +64,32 @@
 
 	let { isCollapsed = false, favoriteLinks = [] }: Props = $props()
 
+	const MAIN_LINK_ICONS: Record<string, any> = { home: Home, runs: Play, schedules: Calendar }
+	let activeWorkspace = $derived($userWorkspaces?.find((w) => w.id === $workspaceStore))
+	let workspaceName = $derived(activeWorkspace?.name ?? $workspaceStore ?? '')
+	// Home is where an operator lands and gets their bearings, so the name stays out there
+	// rather than waiting for a hover. Everywhere else the button is the page's own chrome
+	// and stays out of the way until asked.
+	let showWorkspaceName = $derived(menuOpen || page.url.pathname === `${base}/`)
+	// Most workspaces never get a colour (it is only ever set explicitly), and this button
+	// floats over page content with nothing else behind it — so the disc falls back to a
+	// neutral rather than vanishing, keeping the glyph on a ground of its own everywhere.
+	// A token rather than a hex so it follows the theme; getContrastTextColor only parses
+	// hex, so it returns undefined here and the glyph keeps its themed `text-hint`.
+	const NEUTRAL_DISC = 'rgb(var(--color-surface-sunken))'
+
 	let mainMenuLinks = $derived(
-		[
-			{ label: 'Home', id: 'home', href: `${base}/`, icon: Home },
-			{ label: 'Runs', id: 'runs', href: `${base}/runs`, icon: Play },
-			{ label: 'Schedules', id: 'schedules', href: `${base}/schedules`, icon: Calendar }
-		].filter(
-			(link) =>
-				link.id === 'home' ||
-				($userWorkspaces &&
-					$workspaceStore &&
-					$userWorkspaces.find((_) => _.id === $workspaceStore)?.operator_settings?.[link.id] ===
-						true)
+		OPERATOR_MAIN_LINKS.map((link) => ({ ...link, icon: MAIN_LINK_ICONS[link.id] })).filter(
+			(link) => link.id === 'home' || filterLink(link)
 		)
 	)
 
-	type SecondMenuLink = { label: string; id: string; href: string }
-	function filterLink(link: SecondMenuLink) {
+	function filterLink(link: OperatorMenuLink) {
 		if (!$userWorkspaces || !$workspaceStore) return false
-		let userWorkspace = $userWorkspaces.find((_) => _.id === $workspaceStore)
-		return userWorkspace?.operator_settings?.[link.id] === true
+		return activeWorkspace?.operator_settings?.[link.id] === true
 	}
-	let secondMenuLinks: SecondMenuLink[] = $derived(
-		[
-			{
-				label: 'Resources',
-				id: 'resources',
-				href: `${base}/resources`
-			},
-			{
-				label: 'Variables',
-				id: 'variables',
-				href: `${base}/variables`
-			},
-			{
-				label: 'Assets',
-				id: 'assets',
-				href: `${base}/assets`
-			},
-			{
-				label: 'Groups',
-				id: 'groups',
-				href: `${base}/groups`
-			},
-			{
-				label: 'Folders',
-				id: 'folders',
-				href: `${base}/folders`
-			},
-			{
-				label: 'Workers',
-				id: 'workers',
-				href: `${base}/workers`
-			},
-			{
-				label: 'Audit logs',
-				id: 'audit_logs',
-				href: `${base}/audit_logs`
-			}
-		].filter(filterLink)
-	)
-	type TriggerMenuLink = SecondMenuLink & { kind: string }
-	let allTriggerLinks: TriggerMenuLink[] = $derived(
-		(
-			[
-				{ label: 'Custom HTTP routes', id: 'triggers', href: `${base}/routes`, kind: 'http' },
-				{
-					label: 'Websocket triggers',
-					id: 'triggers',
-					href: `${base}/websocket_triggers`,
-					kind: 'ws'
-				},
-				{
-					label: 'Postgres triggers',
-					id: 'triggers',
-					href: `${base}/postgres_triggers`,
-					kind: 'postgres'
-				},
-				{ label: 'Kafka triggers', id: 'triggers', href: `${base}/kafka_triggers`, kind: 'kafka' },
-				{ label: 'NATS triggers', id: 'triggers', href: `${base}/nats_triggers`, kind: 'nats' },
-				{ label: 'SQS triggers', id: 'triggers', href: `${base}/sqs_triggers`, kind: 'sqs' },
-				{
-					label: 'GCP Pub/Sub triggers',
-					id: 'triggers',
-					href: `${base}/gcp_triggers`,
-					kind: 'gcp'
-				},
-				{
-					label: 'Azure Event Grid triggers',
-					id: 'triggers',
-					href: `${base}/azure_triggers`,
-					kind: 'azure'
-				},
-				{ label: 'MQTT triggers', id: 'triggers', href: `${base}/mqtt_triggers`, kind: 'mqtt' },
-				{ label: 'AMQP triggers', id: 'triggers', href: `${base}/amqp_triggers`, kind: 'amqp' },
-				{ label: 'Email triggers', id: 'triggers', href: `${base}/email_triggers`, kind: 'email' }
-			] as TriggerMenuLink[]
-		).filter(filterLink)
-	)
+	let secondMenuLinks: OperatorMenuLink[] = $derived(OPERATOR_SECONDARY_LINKS.filter(filterLink))
+	let allTriggerLinks: OperatorTriggerLink[] = $derived(OPERATOR_TRIGGER_LINKS.filter(filterLink))
 	let secondMenuTriggerLinks = $derived(
 		allTriggerLinks.filter((link) => $usedTriggerKinds.includes(link.kind))
 	)
@@ -164,24 +105,36 @@
 			placement="bottom-start"
 			openOnHover
 			usePointerDownOutside
+			submenuSafe
+			bind:open={menuOpen}
 			on:close={() => (showExtraTriggers = false)}
 		>
 			{#snippet triggr({ trigger, pinned })}
-				<MenuButton
-					class={twMerge(
-						'!text-xs bg-surface !pl-3.5 !pr-2 !w-auto',
-						// A hover-opened menu leaves the button plain once the pointer moves on;
-						// keeping the tint is what tells you the click pinned it.
-						pinned ? sidebarClasses.selectedBg : ''
-					)}
-					icon={MenuIcon}
-					isCollapsed={false}
-					lightMode
-					label={undefined}
-					{trigger}
-				/>
+				{@const iconColor = getContrastTextColor($workspaceColor)}
+				<!-- Ground for the expanded state only: the button floats over page content and its
+				     hover tint is translucent, so the name would otherwise sit on whatever is
+				     underneath. Icon-only needs none — the disc is always drawn. -->
+				<div class="flex rounded-md {showWorkspaceName ? 'bg-surface' : ''}">
+					<MenuButton
+						class="!text-xs"
+						buttonClass={twMerge(
+							'!pl-3.5 !pr-2 !w-auto',
+							// A hover-opened menu leaves the button plain once the pointer moves on;
+							// keeping the tint is what tells you the click pinned it.
+							pinned ? sidebarClasses.selectedBg : ''
+						)}
+						icon={MenuIcon}
+						isCollapsed={false}
+						lightMode
+						color={$workspaceColor ?? NEUTRAL_DISC}
+						iconProps={iconColor ? { style: `color: ${iconColor}` } : undefined}
+						label={showWorkspaceName ? workspaceName : undefined}
+						ariaLabel={workspaceName ? `Menu — ${workspaceName}` : 'Menu'}
+						{trigger}
+					/>
+				</div>
 			{/snippet}
-			{#snippet children({ item })}
+			{#snippet children({ item, builders })}
 				<div class="w-full max-w-full">
 					{#each favoriteLinks ?? [] as favorite (favorite.href)}
 						<MenuItem
@@ -274,21 +227,7 @@
 							{/if}
 							Switch theme
 						</MenuItem>
-						<MenuItem
-							href="{base}/user/workspaces"
-							onClick={() => clearWorkspaceFromStorage()}
-							lightMode
-							class={twMerge(
-								'flex gap-3.5 px-2 py-2',
-								'transition-colors',
-								sidebarClasses.text,
-								'data-[highlighted]:bg-surface-hover data-[highlighted]:text-primary'
-							)}
-							{item}
-						>
-							<Building size={14} />
-							All workspaces
-						</MenuItem>
+						<SwitchWorkspaceSubmenu {builders} {item} />
 
 						{#if $superadmin}
 							<MenuItem
@@ -321,7 +260,7 @@
 						</MenuItem>
 					</div>
 					<div role="none">
-						{#snippet renderSecondMenuLinks(menuLinks: SecondMenuLink[])}
+						{#snippet renderSecondMenuLinks(menuLinks: OperatorMenuLink[])}
 							{#each menuLinks as menuLink (menuLink.href ?? menuLink.label)}
 								<MenuItem
 									href={menuLink.href}
